requirements.txt
flask==2.0.1
werkzeug==2.0.3
requests==2.26.0
python-dotenv==0.19.0
Flask-Caching==2.0.2
flask-cors==3.0.10